feat: реорганизация проекта и запуск через Docker
- вынес код по feature-папкам (Configuration, Core, Features, Models)
- добавлен volume для SQLite (`/data/kudos.sqlite`)
- экспорт CSV складывается в `./exports` (монтируется в контейнер)
- исправлены пути к БД (теперь абсолютные)
- SessionStore и TelegramService вынесены в отдельные файлы и сделаны public
- обновлён Dockerfile (runtime с curl) и docker-compose.yml (service kudos-bot)
