+++ Package.swift
@@ -1,23 +1,36 @@
-// swift-tools-version:5.10
+// swift-tools-version:5.9
 import PackageDescription
 
 let package = Package(
-    name: "kudos-vapor",
+    name: "ChatBot",
     platforms: [
         .macOS(.v13)
     ],
+    products: [
+        // исполняемый продукт (binary) и библиотека для переиспользования
+        .executable(name: "Run", targets: ["Run"]),
+        .library(name: "App", targets: ["App"])
+    ],
     dependencies: [
-        .package(url: "https://github.com/vapor/vapor.git", from: "4.90.0"),
-        .package(url: "https://github.com/vapor/fluent.git", from: "4.10.0"),
-        .package(url: "https://github.com/vapor/fluent-sqlite-driver.git", from: "4.2.0")
+        .package(url: "https://github.com/vapor/vapor.git", from: "4.92.0"),
+        .package(url: "https://github.com/vapor/fluent.git", from: "4.9.0"),
+        .package(url: "https://github.com/vapor/fluent-sqlite-driver.git", from: "4.5.0")
     ],
     targets: [
-        .executableTarget(
-            name: "Run",
+        .target(
+            name: "App",
             dependencies: [
                 .product(name: "Vapor", package: "vapor"),
                 .product(name: "Fluent", package: "fluent"),
-                .product(name: "FluentSQLiteDriver", package: "fluent-sqlite-driver"),
+                .product(name: "FluentSQLiteDriver", package: "fluent-sqlite-driver")
+            ],
+            path: "Sources/App"
+        ),
+        .executableTarget(
+            name: "Run",
+            dependencies: [
+                .target(name: "App"),
+                .product(name: "Vapor", package: "vapor") // для Application/Environment
             ],
             path: "Sources/Run"
         )
