+++ docker-compose.yml
@@ -7,6 +7,8 @@ services:
     volumes:
       - kudos_db:/data
       - ./exports:/exports
+    ports:
+      - "8080:8080"
     restart: unless-stopped
     healthcheck:
       test: ["CMD", "curl", "-fsS", "http://127.0.0.1:8080/health"]
